# Cranelift: do not check in generated ISLE code; regenerate on every compile.