--- a/ci/ensure_deterministic_build.sh
+++ b/ci/ensure_deterministic_build.sh
@@ -3,6 +3,10 @@
 # This script makes sure that the meta crate deterministically generate files
 # with a high probability.
 # The current directory must be set to the repository's root.
+#
+# We set SKIP_ISLE=1 to skip ISLE generation, because it depends on files
+# in-tree (cranelift/codegen/.../*.isle) but these are not available when we
+# switch to different working directories during this test.
 
 set -e
 
@@ -11,7 +15,7 @@ BUILD_SCRIPT=$(find -wholename "./target/debug/build/cranelift-codegen-*/build-s
 # First, run the script to generate a reference comparison.
 rm -rf /tmp/reference
 mkdir /tmp/reference
-OUT_DIR=/tmp/reference TARGET=x86_64 CARGO_PKG_VERSION=0.1.0 CARGO_MANIFEST_DIR=/tmp $BUILD_SCRIPT
+SKIP_ISLE=1 OUT_DIR=/tmp/reference TARGET=x86_64 CARGO_PKG_VERSION=0.1.0 CARGO_MANIFEST_DIR=/tmp $BUILD_SCRIPT
 
 # To make sure the build script doesn't depend on the current directory, we'll
 # change the current working directory on every iteration. Make this easy to
@@ -36,6 +40,6 @@ do
 
     rm -rf /tmp/try
     mkdir /tmp/try
-    OUT_DIR=/tmp/try TARGET=x86_64 CARGO_PKG_VERSION=0.1.0 CARGO_MANIFEST_DIR=/tmp/src$i $BUILD_SCRIPT
+    SKIP_ISLE=1 OUT_DIR=/tmp/try TARGET=x86_64 CARGO_PKG_VERSION=0.1.0 CARGO_MANIFEST_DIR=/tmp/src$i $BUILD_SCRIPT
     diff -qr /tmp/reference /tmp/try
 done