--- a/cranelift/codegen/meta/src/cdsl/operands.rs
+++ b/cranelift/codegen/meta/src/cdsl/operands.rs
@@ -113,7 +113,6 @@ pub(crate) enum OperandKindFields {
 
 impl OperandKindFields {
     /// Return the [EnumValues] for this field if it is an `enum`.
-    #[cfg(feature = "rebuild-isle")]
     pub(crate) fn enum_values(&self) -> Option<&EnumValues> {
         match self {
             OperandKindFields::ImmEnum(map) => Some(map),