--- a/cranelift/codegen/meta/src/gen_inst.rs
+++ b/cranelift/codegen/meta/src/gen_inst.rs
@@ -1,6 +1,5 @@
 //! Generate instruction data (including opcodes, formats, builders, etc.).
 use std::fmt;
-use std::path::Path;
 
 use cranelift_codegen_shared::constant_hash;
 
@@ -1085,7 +1084,6 @@ fn gen_inst_builder(inst: &Instruction, format: &InstructionFormat, fmt: &mut Fo
     fmtln!(fmt, "}")
 }
 
-#[cfg(feature = "rebuild-isle")]
 fn gen_isle(formats: &[&InstructionFormat], instructions: &AllInstructions, fmt: &mut Formatter) {
     use std::collections::{BTreeMap, BTreeSet};
     use std::fmt::Write;
@@ -1341,7 +1339,6 @@ fn gen_isle(formats: &[&InstructionFormat], instructions: &AllInstructions, fmt:
 }
 
 /// Generate an `enum` immediate in ISLE.
-#[cfg(feature = "rebuild-isle")]
 fn gen_isle_enum(name: &str, mut variants: Vec<&str>, fmt: &mut Formatter) {
     variants.sort();
     let prefix = format!(";;;; Enumerated Immediate: {} ", name);
@@ -1407,7 +1404,7 @@ pub(crate) fn generate(
     inst_builder_filename: &str,
     isle_filename: &str,
     out_dir: &str,
-    crate_dir: &Path,
+    isle_dir: &str,
 ) -> Result<(), error::Error> {
     // Opcodes.
     let mut fmt = Formatter::new();
@@ -1424,18 +1421,9 @@ pub(crate) fn generate(
     fmt.update_file(opcode_filename, out_dir)?;
 
     // ISLE DSL.
-    #[cfg(feature = "rebuild-isle")]
-    {
-        let mut fmt = Formatter::new();
-        gen_isle(&formats, all_inst, &mut fmt);
-        let crate_src_dir = crate_dir.join("src");
-        fmt.update_file(isle_filename, &crate_src_dir.display().to_string())?;
-    }
-    #[cfg(not(feature = "rebuild-isle"))]
-    {
-        // Silence unused variable warnings.
-        let _ = (isle_filename, crate_dir);
-    }
+    let mut fmt = Formatter::new();
+    gen_isle(&formats, all_inst, &mut fmt);
+    fmt.update_file(isle_filename, isle_dir)?;
 
     // Instruction builder.
     let mut fmt = Formatter::new();