--- a/cranelift/codegen/src/isa/aarch64/lower/isle/generated_code.manifest
+++ b/cranelift/codegen/src/isa/aarch64/lower/isle/generated_code.manifest
@@ -1,4 +1,0 @@
-src/clif.isle 443b34b797fc8ace
-src/prelude.isle e6c91b0115343ab9
-src/isa/aarch64/inst.isle 21a43af20be377d2
-src/isa/aarch64/lower.isle 75ad8450963e3829