--- a/cranelift/codegen/src/isa/s390x/lower/isle/generated_code.manifest
+++ b/cranelift/codegen/src/isa/s390x/lower/isle/generated_code.manifest
@@ -1,4 +1,0 @@
-src/clif.isle 443b34b797fc8ace
-src/prelude.isle e6c91b0115343ab9
-src/isa/s390x/inst.isle 36c2500563cdd4e6
-src/isa/s390x/lower.isle e5c946ab8a265b77