--- a/cranelift/codegen/src/isa/x64/lower/isle/generated_code.manifest
+++ b/cranelift/codegen/src/isa/x64/lower/isle/generated_code.manifest
@@ -1,4 +1,0 @@
-src/clif.isle 443b34b797fc8ace
-src/prelude.isle e6c91b0115343ab9
-src/isa/x64/inst.isle 833710d359126637
-src/isa/x64/lower.isle 4c567e9157f84afb