… .github/workflows/main.yml
… Cargo.lock
… cranelift/codegen/Cargo.toml
… cranelift/codegen/meta/Cargo.toml
… cranelift/codegen/src/clif.isle
… cranelift/codegen/src/isa/aarch64/lower/isle/generated_code.rs
… cranelift/codegen/src/isa/s390x/lower/isle/generated_code.rs
… cranelift/codegen/src/isa/x64/lower/isle/generated_code.rs
… cranelift/isle/isle/Cargo.toml
… cranelift/isle/isle/src/error.rs
… cranelift/isle/isle/src/error_miette.rs
… cranelift/isle/isle/src/ir.rs
… cranelift/isle/isle/src/lexer.rs
… cranelift/isle/isle/src/lib.rs
… cranelift/isle/isle/src/log.rs
… cranelift/isle/isle/src/parser.rs
… cranelift/isle/isle/src/sema.rs
… cranelift/isle/isle/src/trie.rs
… cranelift/isle/isle/tests/run_tests.rs
… cranelift/isle/islec/Cargo.toml
… cranelift/isle/islec/src/main.rs
… scripts/force-rebuild-isle.sh