@cfallin 2022-05-12
Thanks! Updated and fixed the deterministic-meta test (previously was skipping ISLE build entirely, now does so again); will merge with your r+.