@bjorn3 2022-05-12
For me cranelift-isle takes 10s to build. The full `cargo check -p cranelift-codegen` takes 25s. `cargo build -p cranelift-codegen` takes 35s and `cargo build -p cranelift-codegen --release` takes 50s. So it is not ideal, but not the biggest problem either. I will see if there are some easy wins to have.