@fitzgen 2022-05-12
> With a default configuration, an ISLE error looks like this during build:
> 
> ```
>   --- stderr
>   Error building ISLE files: ISLE errors:
> 
>   parse error: Unexpected token Symbol("asdf")
> 
>   To see a more detailed error report, run:
> 
>       $ cargo check -p cranelift-codegen --features isle-errors
> ```

Any chance we could still have a source position in there when the `isle-errors` feature is not enabled?