@cfallin 2022-05-12
> Any chance we could still have a source position in there when the `isle-errors` feature is not enabled?

Sure! #4151